+++ refena/lib/src/provider/types/view_family_provider.dart
@@ -10,7 +10,10 @@ typedef ViewFamilyBuilder<T, P> = T Function(WatchableRef ref, P param);
 /// Similar to [ViewProvider] but with a parameter.
 /// It is essentially a syntax sugar for ViewProvider<Map<P, T>>.
 class ViewFamilyProvider<T, P>
-    extends BaseProvider<FamilyNotifier<T, P, ViewProvider<T>>, Map<P, T>> {
+    extends BaseProvider<FamilyNotifier<T, P, ViewProvider<T>>, Map<P, T>>
+    implements
+        RebuildableProvider<FamilyNotifier<T, P, ViewProvider<T>>, Map<P, T>,
+            void> {
   final ViewFamilyBuilder<T, P> _builder;
   final String Function(T state)? _describeState;